move Template to its own file
