+++ lib/view_component/base.rb
@@ -11,6 +11,7 @@ require "view_component/inline_template"
 require "view_component/preview"
 require "view_component/slotable"
 require "view_component/slotable_default"
+require "view_component/template"
 require "view_component/translatable"
 require "view_component/with_content_helper"
 require "view_component/use_helpers"
